change page to async without reload button
