+++ apps/cow-amm-deployer/src/app/[userId]/amms/[id]/withdraw/(components)/WithdrawForm.tsx
@@ -21,7 +21,13 @@ import {
 } from "#/lib/transactionFactory";
 import { ChainId } from "#/utils/chainsPublicClients";
 
-export function WithdrawForm({ cowAmm }: { cowAmm: ICowAmm }) {
+export function WithdrawForm({
+  cowAmm,
+  userId,
+}: {
+  cowAmm: ICowAmm;
+  userId: string;
+}) {
   const form = useForm<typeof ammWithdrawSchema._type>({
     // @ts-ignore
     resolver: zodResolver(ammWithdrawSchema),
@@ -38,11 +44,11 @@ export function WithdrawForm({ cowAmm }: { cowAmm: ICowAmm }) {
   const onSubmit = async (data: typeof ammWithdrawSchema._type) => {
     const amount0 = parseUnits(
       String((Number(cowAmm.token0.balance) * data.withdrawPct) / 100),
-      cowAmm.token0.decimals,
+      cowAmm.token0.decimals
     );
     const amount1 = parseUnits(
       String((Number(cowAmm.token1.balance) * data.withdrawPct) / 100),
-      cowAmm.token1.decimals,
+      cowAmm.token1.decimals
     );
     const txArgs = {
       type: TRANSACTION_TYPES.WITHDRAW,
@@ -53,7 +59,7 @@ export function WithdrawForm({ cowAmm }: { cowAmm: ICowAmm }) {
     } as withdrawCowAMMargs;
     try {
       await sendTransactions([txArgs]);
-      router.push(`/amms/${cowAmm.id}`);
+      router.push(`${userId}/amms/${cowAmm.id}`);
     } catch {
       toast({
         title: `Transaction failed`,
@@ -116,7 +122,7 @@ export function WithdrawForm({ cowAmm }: { cowAmm: ICowAmm }) {
                     <span className="font-semibold">
                       {formatNumber(
                         (Number(token.balance) * withdrawPct) / 100,
-                        4,
+                        4
                       )}{" "}
                       {token.symbol}
                     </span>
@@ -124,7 +130,7 @@ export function WithdrawForm({ cowAmm }: { cowAmm: ICowAmm }) {
                       $
                       {formatNumber(
                         (Number(token.usdValue) * withdrawPct) / 100,
-                        4,
+                        4
                       )}
                     </span>
                   </div>
