+++ apps/cow-amm-deployer/src/components/TokenSelect.tsx
@@ -48,7 +48,7 @@ export function TokenSelect({
     try {
       const importedToken = await fetchTokenInfo(
         search as Address,
-        chainId as ChainId,
+        chainId as ChainId
       );
       handleSelectToken(importedToken);
       addImportedToken(importedToken, chainId as ChainId);
@@ -118,11 +118,13 @@ export function TokenSelect({
                   onSelect={() => handleSelectToken(token)}
                 >
                   <TokenInfo
-                    token={{
-                      address: token.address as Address,
-                      symbol: token.symbol,
-                      decimals: token.decimals,
-                    }}
+                    token={
+                      {
+                        address: token.address,
+                        symbol: token.symbol,
+                        decimals: token.decimals,
+                      } as IToken
+                    }
                   />
                 </CommandItem>
               ))}
